Drop the hidden visibility from DebugHandlerBase for now.
If a class has hidden visibility all derived classes and all classes
that have it as a member must have hidden visibility too. That may
be fixable here but requires changes to quite a lot of debug info
classes.

This is also one of the things that GCC enforces aggressively while
clang ignores it, making testing more annoying than necessary.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@260529 91177308-0d34-0410-b5e6-96231b3b80d8