[SROA] Clean up some really weird code, no functionality changed.
We already have the instruction extracted into 'I', just cast that to
a store the way we do for loads. Also, we don't enter the if unless SI
is non-null, so don't test it again for null.

I'm pretty sure the entire test there can be nuked, but this is just the
trivial cleanup.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@263112 91177308-0d34-0410-b5e6-96231b3b80d8