don't hardcode the name of the llc checks script
We lose the 'utils' directory name in our advertising line with
this change. We could retain that, but I don't see the point.
This removes a dependency for making the script apply to more than
'llc'. Ie, we'll want to change the script name if it works with 
opt/clang too.



git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@264310 91177308-0d34-0410-b5e6-96231b3b80d8