Improve CHECK-NOT robustness of dllexport tests
This changes some dllexport tests, to verify that some symbols that
should not be exported are not, in a way that improves the robustness
of CHECK-SAME interaction with CHECK-NOT.

We plan to enable dllimport/dllexport support for the PS4, and these
changes are for points we noticed in our internal testing.

Patch by Warren Ristow!

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@265106 91177308-0d34-0410-b5e6-96231b3b80d8